1.Add bottom menu. 2.Add async task loader. 3.Work with activities. 4.Work with layouts.

+++ app/src/main/java/com/masliaiev/filmspace/MainActivity.java
@@ -1,22 +1,32 @@
 package com.masliaiev.filmspace;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
+import android.util.DisplayMetrics;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.CompoundButton;
-import android.widget.RadioGroup;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.android.material.navigation.NavigationBarView;
+import com.masliaiev.filmspace.adapters.MovieAdapter;
 import com.masliaiev.filmspace.data.MainViewModel;
 import com.masliaiev.filmspace.data.Movie;
 import com.masliaiev.filmspace.utils.JSONUtils;
@@ -24,46 +34,80 @@ import com.masliaiev.filmspace.utils.NetworkUtils;
 
 import org.json.JSONObject;
 
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
 
     private Switch switchSort;
     private RecyclerView recyclerViewPosters;
     private MovieAdapter movieAdapter;
     private TextView textViewPopularity;
     private TextView textViewTopRated;
+    private BottomNavigationView bottomNavigationView;
+    private ProgressBar progressBarLoading;
 
     private MainViewModel viewModel;
+
+    private static final int LOADER_ID = 133;
+    private LoaderManager loaderManager;
+
+    private static int page = 1;
+    private static int methodOfSort;
+    private static boolean isLoading = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.hide();
+        }
+        loaderManager = LoaderManager.getInstance(this);
+        bottomNavigationView = findViewById(R.id.bottomNavigationView);
+        Menu menu = bottomNavigationView.getMenu();
+        menu.findItem(R.id.bottomHome).setIcon(R.drawable.home_white);
+        bottomNavigationView.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                switch (item.getItemId()) {
+                    case R.id.bottomHome:
+                        break;
+                    case R.id.bottomFavourites:
+                        Intent intentFavourites = new Intent(MainActivity.this, FavouriteActivity.class);
+                        startActivity(intentFavourites);
+                        overridePendingTransition(0,0);
+                        finish();
+                        break;
+                    case R.id.bottomRandom:
+                        Intent intentRandom = new Intent(MainActivity.this, RandomMovieActivity.class);
+                        startActivity(intentRandom);
+                        overridePendingTransition(0,0);
+                        finish();
+                        break;
+                    default:
+                        throw new IllegalStateException("Unexpected value: " + item.getItemId());
+                }
+                return false;
+            }
+        });
         viewModel = new  ViewModelProvider(this).get(MainViewModel.class);
+        progressBarLoading = findViewById(R.id.progressBarLoading);
         switchSort = findViewById(R.id.switchSort);
         textViewPopularity = findViewById(R.id.textViewPopularity);
         textViewTopRated = findViewById(R.id.textViewTopRated);
         recyclerViewPosters = findViewById(R.id.recyclerViewPosters);
-        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, 2));
+        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, getColumnCount()));
         movieAdapter = new MovieAdapter();
         recyclerViewPosters.setAdapter(movieAdapter);
         switchSort.setChecked(true);
         switchSort.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                int methodOfSort;
-                if (isChecked) {
-                    methodOfSort = NetworkUtils.TOP_RATED;
-                    textViewPopularity.setTextColor(getResources().getColor(R.color.white));
-                    textViewTopRated.setTextColor(getResources().getColor(R.color.teal_200));
-                } else {
-                    methodOfSort = NetworkUtils.POPULARITY;
-                    textViewPopularity.setTextColor(getResources().getColor(R.color.teal_200));
-                    textViewTopRated.setTextColor(getResources().getColor(R.color.white));
-                }
-                downloadData(methodOfSort, 1);
+                page = 1;
+                setMethodOfSort(isChecked);
             }
         });
         switchSort.setChecked(false);
@@ -79,7 +123,9 @@ public class MainActivity extends AppCompatActivity {
         movieAdapter.setOnReachEndListener(new MovieAdapter.OnReachEndListener() {
             @Override
             public void onReachEnd() {
-                Toast.makeText(MainActivity.this, "End", Toast.LENGTH_SHORT).show();
+                if (!isLoading) {
+                    downloadData(methodOfSort, page);
+                }
             }
         });
 
@@ -87,29 +133,87 @@ public class MainActivity extends AppCompatActivity {
         moviesFromLiveData.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                movieAdapter.setMovies(movies);
+                if (page == 1) {
+                    movieAdapter.setMovies(movies);
+                }
             }
         });
 
     }
 
     private void downloadData (int methodOfSort, int page) {
-        JSONObject jsonObject = NetworkUtils.getJSONFromNetwork(methodOfSort, page);
-        ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(jsonObject);
-        movieAdapter.setMovies(movies);
-        if (!movies.isEmpty()) {
-            viewModel.deleteAllMovies();
-            for (Movie movie : movies) {
-                viewModel.insertMovie(movie);
-            }
-        }
+        URL url = NetworkUtils.buildURL(methodOfSort, page);
+        Bundle bundle = new Bundle();
+        bundle.putString("url", url.toString());
+        loaderManager.restartLoader(LOADER_ID, bundle, this);
     }
 
     public void onClickSetPopularity(View view) {
+        setMethodOfSort(false);
         switchSort.setChecked(false);
     }
 
     public void onClickSetTopRated(View view) {
+        setMethodOfSort(true);
         switchSort.setChecked(true);
+    }
+
+    private void setMethodOfSort (boolean isTopRated) {
+        if (isTopRated) {
+            methodOfSort = NetworkUtils.TOP_RATED;
+            textViewPopularity.setTextColor(getResources().getColor(R.color.white));
+            textViewTopRated.setTextColor(getResources().getColor(R.color.teal_200));
+        } else {
+            methodOfSort = NetworkUtils.POPULARITY;
+            textViewPopularity.setTextColor(getResources().getColor(R.color.teal_200));
+            textViewTopRated.setTextColor(getResources().getColor(R.color.white));
+        }
+        downloadData(methodOfSort, page);
+    }
+
+    private int getColumnCount () {
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        int width = (int) (displayMetrics.widthPixels / displayMetrics.density);
+        return width / 185 > 2 ? width / 185 : 2;
+    }
+
+    @NonNull
+    @Override
+    public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle args) {
+        NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this, args);
+        jsonLoader.setOnStartLoadingListener(new NetworkUtils.JSONLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
+        return jsonLoader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
+        ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(data);
+        if (movies != null && !movies.isEmpty()) {
+            if (page == 1) {
+                viewModel.deleteAllMovies();
+                movieAdapter.clear();
+            }
+            for (Movie movie : movies) {
+                viewModel.insertMovie(movie);
+            }
+            movieAdapter.addMovies(movies);
+            page++;
+        }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
+        loaderManager.destroyLoader(LOADER_ID);
+
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<JSONObject> loader) {
+
     }
 }
